refactor: split EventMenu into focused classes under UI/Menus/Events/
- Extract browse, search, and filter logic into EventBrowseMenu.cs
- Extract create event flow into EventCreateMenu.cs
- Extract my events, edit, and cancel into EventManageMenu.cs
- Move EventDetailMenu.cs into Events/ subfolder
- Update namespaces to EventPlatform.UI.Menus.Events
- Update MainMenu to use new split classes
- Delete EventMenu.cs
